Gui branch

+++ Engine/src/engine/api/dto/InstructionPeek.java
@@ -5,7 +5,7 @@ public record InstructionPeek(
         String stringRepresentation,
         String label,
         boolean isSynthetic,
-        int cycles,
+        long cycles,
         // which instruction was this expanded from
         InstructionPeek expandedFrom, // can be null
         int lineId
